chore: Add test for --exclude-base-image-vulns
This functionality is used enough to necessitate adding another
acceptance test. It is "most-used-path" test for this option and it
doesn't test combinations with other options, like --json.

In order to reduce duplication, fixtures are shared with similar "docker
remediation advice" test. Docker plugin response stub growing quite big,
it is extracted into a fixture file, which is also reused among the two
tests.