fix: switch to openjdk in java-related docker images
Snyk CLI is distributed as an npm package, standalone binaries an couple of specific docker images. Java related ecosystems (Maven, Gradle, SBT) require JDK to be present. Because of the license change, Oracle JDK is not available anymore. In order to keep relate  docker images working, another version of JDK has to be used.