feat: log detected orphaned gradle files + test
If a gralde manifest was detected early on but no dependencies were returned at all for it
warn the user to look further as this file may not be included in a root level gradle file.

Thus it could be orhpaned. It could have also failed in which case an error will pop up.