feat: Auto detect user instructions from image
Added auto detecting user instructions from container image config, so from now on customer won't need to provide a Dockerfile and will get this info for free. This can be seen on the individual vulnerability card as "Dockerfile instruction" and is also used to provide a filter on the project page. Bumping the plugin releases that feature for `snyk container monitor`.