refactor: Remove use of options.iacDirFiles
We were only using these for the display output in the test/index.ts
file, this change replaces this usage with a local variable. Now the
only place using the iacDirFiles property is the TestOptions and the
legacy remote execution test flow.

One change from this PR is that we will now always use the file basename
in the CLI output rather than the basename for directory scanning and
the full path for the single file scanning.