fix: exit code when IaC scans are empty
According to
https://support.snyk.io/hc/en-us/articles/360003812578#UUID-c88e66cf-431c-9ab1-d388-a8f82991c6e0,
the Snyk CLI should return 3 when there were no supported files in a
scan. This allows user to distinguish this situation from a complete
scan, and from when errors encountered during the scan.

We were actually returning 2, which is also returned when errors are
encountered during a scan.