perf: Fix OOM issues caused by group-issues
Group issues mechanism was extending the "from" and "name" arrays in
the current vuln it was iterating on, and only then assigning it to
acc[vuln.id]. This meant that for every vulns that needs to be grouped,
an array larger by 1 (of "from" and of "name") is added to the vuln.
In large vulns results this may cause may OOM issues,
because of data duplication.
This fix changes the the mechanism so that when acc[vuln.id] exists,
it adds the current vuln's 'from' and 'name' values to the existing
acc[vuln.id].from and name arrays.