feat: Add TF Variable support for nested directories
This commit adds support for dereferencing variables and finding issues due to them in the path provided by the user, but also in nested directories.

We send TF files per directory to the parser, where the variables are extracted and merged all together, sharing a variable context.
Then, the dereferenced json file content is sent back to the CLI where it gets evaluated, etc and returns issues to the user as normal.
This is still under the `iacTerraformVarSupport` flag.

CFG-1483