fix: Increase delay between inputs on some acceptance tests
Our prompt library inquirer only receives data from stdin after it has
been initialised. Therefore we can lose some buffered data if we send it
too fast, this is particularly noticeable if there are multiple inquirer
prompts in a row. There is no known way to check if the child process
that is under test is ready to receive input as node will swallow
anything given to stdin if it has nowhere to send it.

This hopefully will decrease the probability that these test fail by
increasing the delay between sending data to the child process.