feat: developer flag to override IaC bundle path
When developing certain Snyk IaC features, it can be useful to override
the OPA WASM bundle with a locally-built version. This commit introduces
and environment variable, SNYK_IAC_BUNDLE_PATH, which is not documented
anywhere. This is intended for Snyk IaC developers, and is not useful
for other users. The user-facing equivalent is custom rule bundles.