fix: Skip unsupported IaC files
This commit will fix an issue where unsupported IaC files were printed in the scan output for users without the iacTerraformVarSupport flag.
This was due to a refactoring, where a check for the valid IaC files was skipped.
This is now added (hardcoded) for the flow that the flag is off.