fix: ignore errors on multiple paths if there were some results
This commit fixes a bug where if we had multiple paths in a scan, and one of them threw an error, the entire scan was failing.
Now we are going to still show success results if there are any, and not fail the entire scan.
The bug affected both the legacy output and the new output.

CFG-1784