chore: remove check dev environment prepare hook
npm's prepare hook is unreliable and changes between npm versions. Currently cli-alert broke as this parent hook now triggers during `npm ci` on a child workspace, even though `npm ci` on a workspace does not install all parent dependencies. In this case, ts-node isn't installed so it can't run check-dev-environment.ts and fails.

I've gone for removing check-dev-environment entirely. We now require Node 16 and whatever npm it comes with so enforcing this manually is simpler. If we need this again, we can introduce it in the `Makefile` with simpler `npm --version` checks without relying on npm hooks.