feat: add `IAC_RULES_CLIENT_URL` to override the URL of the rules client in snyk-iac-test
This is a "hidden" env var to override the url of the rulesClient that is being used later in snyk-iac-test, to create a client for opa-rules-distribution.
With this flag, we will be able to test bundles in development, while going through the opa-rules-distribution automatic download process.

To use locally:
``` SNYK_IAC_RULES_CLIENT_URL=https://something.com snyk-dev iac test test/fixtures/iac/terraform/sg_open_ssh.tf
```

chore: bump snyk-iac-test to 0.42.0